#155 refactor: 유효성 검증 util 분리
